Re-work mock_time return to be an instance rather than a class
The removes the need for __new__ complexities, and so makes typing simpler, but does change the repr.

As a result, many tests were simplified to use the mock as a fixture/replace-return rather than using the mocked `time` directly.
This meant a bunch of ugly casts and local imports could be removed 🎉.

The two reproducer tests added for calling a `mock_time()` with parameters were removed, this edge case functionality loss should be added to the release notes.